Move block_details and block_sideband implementations to its own file.
